requirements.txt
aws-cdk-lib==2.143.0
constructs>=10.0.5
cdklabs.generative-ai-cdk-constructs==0.1.158
PyPaperBot